a somewhat working solution

+++ ProcessStreamer/HelperExtensions.cs
@@ -6,9 +6,9 @@ namespace ProcessStreamer
 		private readonly static DateTime unixEpoch =
 			new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
 
-		public static double ToUnixTimeSeconds(this DateTime time)
+		public static int ToUnixTimeSeconds(this DateTime time)
 		{
-			return time.Subtract(unixEpoch).TotalSeconds;
+			return (int)time.Subtract(unixEpoch).TotalSeconds;
 		}
     }
 }
